% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_missForest.R
\name{autotune_missForest}
\alias{autotune_missForest}
\title{Perform imputation using missForest form missForest package.}
\usage{
autotune_missForest(
  df,
  percent_of_missing,
  cores = NULL,
  ntree_set = c(100, 200, 500, 1000),
  mtry_set = NULL,
  parallel = TRUE,
  turn_off_parallel = FALSE,
  col_0_1 = FALSE,
  optimize = TRUE,
  ntree = 100,
  mtry = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without  target column.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{cores}{integer.  Number of threads used by parallel calculations. By default approximately half of available CPU cores.}

\item{ntree_set}{integer vector. Vector contains numbers of tree for grid search.}

\item{mtry_set}{integer vector. Vector contains numbers of variables randomly sampled at each split.}

\item{parallel}{logical. If TRUE parallel calculation is using.}

\item{turn_off_parallel}{logical. If TRUE parallel backend is turn off after imputation.}

\item{col_0_1}{decide if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False.}

\item{optimize}{optimize inside function}

\item{ntree}{ntree from missForest function}

\item{mtry}{mtry form missforest function}

\item{verbose}{If FALSE funtion didn't print on console.}
}
\value{
Return data.frame with imputed values.
}
\description{
Function use missForest package for data imputation. Function use OBBerror (more in missForest documentation) to perform grid search.
Best parameters to imputation are chosen form given sets. Imputation use parallel calculation by default and use existing parallel backend if it's possible.
If not function starts new parallel backend. !!!! Function doesn't turn off parallel backend by default after imputation. !!!!
}
