% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_rMIDAS.R
\name{autotune_rMIDAS}
\alias{autotune_rMIDAS}
\title{Multiple Imputation using Denoising Autoencoders}
\usage{
autotune_rMIDAS(
  data,
  col_type,
  path,
  type,
  training_epchos = 10L,
  layer_structure = c(256, 256, 256),
  seed = 123,
  learn_rate = 4e-04
)
}
\arguments{
\item{data}{Data frame with missing values}

\item{col_type}{Columns type in data frame}

\item{path}{haracter string, path to python binary if type == "auto", path to virtualenv if type == "virtualenv", or the name of a Conda environment if type=="condaenv". More information aboute Pythone enviroment can be found in Detalis.}

\item{type}{character string, one of 'auto' (for python binary),'virtualenv', or 'condaenv. More information aboute Pythone enviroment can be found in Detalis.}

\item{training_epchos}{An integer, indicating the number of forward passes to conduct when running the model.}

\item{layer_structure}{A vector of integers, The number of nodes in each layer of the network (default = c(256, 256, 256), denoting a three-layer network with 256 nodes per layer). Larger networks can learn more complex data structures but require longer training and are more prone to overfitting.}

\item{seed}{Random seed}

\item{learn_rate}{A number, the learning rate γ (default = 0.0001), which controls the size of the weight adjustment in each training epoch. In general, higher values reduce training time at the expense of less accurate results.}
}
\description{
The function creates an easy interface for rMIDAS package. More information about specific can be found here http://commonmark.org/help.
}
\details{
The function requires a Python environment created with package reticulate. The environment has to be created before function call and wouldn't be close by function. An example of Python background can be found in Example sections.
}
\examples{
 #Creating Python enviroment using conda


 reticulate::conda_create('impute')
 reticulate::conda_install('impute','scikit-learn')
 rMIDAS::set_python_env('impute',type="condaenv")

 autotune_rMIDAS(data,col_type,'impute','condaenv')

 # Can be change to any difrent  Python enviroment


}
\author{
Thomas Robinson ,, Ranjit Lall ORCID iD, Alex Stenlake [
}
