% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_VIM_Irmi.R
\name{autotune_VIM_Irmi}
\alias{autotune_VIM_Irmi}
\title{Perform imputation using VIM package and irmi function}
\usage{
autotune_VIM_Irmi(
  df,
  percent_of_missing,
  eps = 5,
  maxit = 100,
  step = FALSE,
  robust = FALSE,
  init.method = "kNN",
  force = FALSE,
  col_0_1 = FALSE,
  out_file = NULL
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without target column.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{eps}{threshold for convergency}

\item{maxit}{maximum number of iterations}

\item{step}{stepwise model selection is applied when the parameter is set to TRUE}

\item{robust}{if TRUE, robust regression methods will be applied (it's impossible to set step=TRUE and robust=TRUE at the same time)}

\item{init.method}{Method for initialization of missing values (kNN or median)}

\item{force}{if TRUE, the algorithm tries to find a solution in any case, possible by using different robust methods automatically. (should be set FALSE for simulation)}

\item{col_0_1}{Decaid if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False. (Works only for returning one dataset).}

\item{out_file}{Output log file location if file already exists log message will be added. If NULL no log will be produced.}
}
\value{
Return one data.frame with imputed values.
}
\description{
Function use IRMI (Iterative robust model-based imputation ) to impute missing data.
}
