% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_VIM_kNN.R
\name{autotune_VIM_kNN}
\alias{autotune_VIM_kNN}
\title{K nearest neighbor imputation using VIM package.}
\usage{
autotune_VIM_kNN(
  df,
  percent_of_missing,
  k = 5,
  numFun = median,
  catFun = maxCat,
  col_0_1 = FALSE,
  optimize = FALSE,
  number_of_try = 10
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without  target column.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{k}{Value of k use if optimize=FALSE}

\item{col_0_1}{decide if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False.}

\item{optimize}{If True function try to optimize k. Its can be pointless in many casys becouse. Evalutaion funtion from VIM ist not vary sutible for this task.
If False chosen k is used.}

\item{number_of_try}{Number of k betwen 5 and one-thired of dataset rows teseted for best k.}
}
\value{
Return data.frame with imputed values.
}
\description{
Function perform kNN function from VIM packge. Only K is tuned by function.
}
