% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA-data.R
\docType{data}
\name{de_res_IFNg_vs_naive}
\alias{de_res_IFNg_vs_naive}
\title{A sample \code{data.frame} containing Differential Expression Analysis, generated
with \code{DESeq2}}
\format{
A \code{data.frame} object
}
\value{
A sample \code{data.frame} object, extracted from \code{DESeq2} results
}
\description{
A sample \code{data.frame} containing Differential Expression Analysis, generated
with \code{DESeq2}
}
\details{
This \code{data.frame} object contains the results of a Differential
Expression Analysis performed on data from the \code{macrophage} package, more
precisely contrasting the counts from naive macrophage to those associated
with IFNg.

The code to create said object can be found in the folder \verb{/inst/scripts} in
the EMMA package, the file is called \code{create_datasets_examples.R}.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
