% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_get_record.R
\name{EMMA_get_record}
\alias{EMMA_get_record}
\title{EMMA_get_record}
\usage{
EMMA_get_record(res)
}
\arguments{
\item{res}{Functional Enrichment Analysis results (enrichResult, gseaResult ...)
generated by \code{EMMA_run()}}
}
\value{
list of metadata recorded during FEA runtime
}
\description{
EMMA_get_record
}
\examples{
data("fea_res", package = "EMMA")
EMMA_get_record(fea_res)
}
