% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_run.R
\name{EMMA_run}
\alias{EMMA_run}
\title{EMMA_run}
\usage{
EMMA_run(expr, envir = parent.frame())
}
\arguments{
\item{expr}{A function call that performs functional enrichment analysis.
The call is captured and executed by EMMA to record analysis parameters and
provenance information}

\item{envir}{An environment in which to evaluate \code{expr}}
}
\value{
Functional enrichment analysis results in the native format
returned by the original \code{expr}
}
\description{
This function executes functional enrichment analysis using existing tools
and captures the associated parameters and provenance information for the
analysis during runtime.
}
\examples{
EMMA_run(enrichGO(gene = rownames(de_res_IFNg_vs_naive), universe = universe, keyType = "ENSEMBL", OrgDb = org.Hs.eg.db, ont = "BP"))
}
