\name{emmadat}
\docType{emmadat}
\title{Sample inpute data for EMMA}
\description{
  This dataset is consist of three matries. \code{K} is the kinship
  matrix between 110 individuals, \code{xs} is the 100 simulated SNPs
  encoded with zeros and ones, and \code{ys} is the 50 simulated
  phenotypes as doubles.
}
\usage{emmadat}
\format{A list consisting of 110 x 110 matrix (K), 100 x 110 matrix
  (xs), and 50 x 110 matrix(ys)
}
\keyword{datasets}