% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA-data.R
\docType{data}
\name{fea_res}
\alias{fea_res}
\title{A sample \code{list} containing Functional Enrichment Analysis results,
generated with \code{gprofiler2}}
\format{
A \code{list}
}
\value{
A sample \code{list} containing the FEA results \code{result} and \code{metadata}.
This results object has the \code{EMMA_record} attribute.
}
\description{
A sample \code{list} containing Functional Enrichment Analysis results,
generated with \code{gprofiler2}
}
\details{
This \code{list} object contains the result table and metadata of the
functional enrichment analysis (FEA) performed on the \code{macrophage} data,
specifically using the \code{gost()} function from the \code{gprofiler2} package, and
wrapped in \code{EMMA_run()}

The code to create said object can be found in the folder \verb{/inst/scripts} in
the EMMA package, the file is called \code{create_datasets_examples.R}.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
