% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMDA_MFA.R
\name{missMDA_MFA}
\alias{missMDA_MFA}
\title{Perform imputation using MFA algorithm.}
\usage{
missMDA_MFA(
  df,
  col_type,
  percent_of_missing,
  random.seed = 123,
  ncp = 2,
  col_0_1 = F
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without target column.}

\item{col_type}{character vector. Vector containing column type names.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{random.seed}{random seed.}

\item{ncp}{Number of dimensions used by algorithm. Default 2.}

\item{col_0_1}{Decaid if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False. (Works only for returning one dataset).}
}
\value{
Return one data.frame with imputed values.
}
\description{
Function use MFA (Multiple Factor Analysis) to impute missing data. Groups are created using original column order and taking as much variable to one group as possible.
This method doesn't allow to Multiple Imputation. Function doesn't use any optimization teaching for ncp param.
}
\details{
MFA requires to select group type but numeric types can only be set as 'c' - centered and 's' - scale to unit variance.
It's impossible to provide these conditions so numeric type is always set as 's'.
}
