% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOPAmelia.R
\name{PipeOpAmelia}
\alias{PipeOpAmelia}
\title{Amelia imputation}
\description{
This class create object implements autotune_Amelia function for use in mlr3 pipelinies. Object can be created with \code{\link{autotune_Amelia}} params.
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpImpute]{mlr3pipelines::PipeOpImpute}} -> \code{Amelia_imputation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PipeOpAmelia$new()}}
\item \href{#method-train_imputer}{\code{PipeOpAmelia$train_imputer()}}
\item \href{#method-impute}{\code{PipeOpAmelia$impute()}}
\item \href{#method-clone}{\code{PipeOpAmelia$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict">}\href{../../mlr3pipelines/html/PipeOp.html#method-predict}{\code{mlr3pipelines::PipeOp$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print">}\href{../../mlr3pipelines/html/PipeOp.html#method-print}{\code{mlr3pipelines::PipeOp$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train">}\href{../../mlr3pipelines/html/PipeOp.html#method-train}{\code{mlr3pipelines::PipeOp$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpAmelia$new(
  id = "imput_Amelia",
  col_0_1 = FALSE,
  polytime = NULL,
  splinetime = NULL,
  intercs = FALSE,
  empir = NULL,
  m = 3,
  parallel = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_imputer"></a>}}
\if{latex}{\out{\hypertarget{method-train_imputer}{}}}
\subsection{Method \code{train_imputer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpAmelia$train_imputer(feature, type, context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-impute"></a>}}
\if{latex}{\out{\hypertarget{method-impute}{}}}
\subsection{Method \code{impute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpAmelia$impute(feature, type, model, context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpAmelia$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
