% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_reuse.R
\name{replace_overimputes}
\alias{replace_overimputes}
\title{Replace overimputes. Used in mice.reuse.}
\usage{
replace_overimputes(data, imp, j, i)
}
\arguments{
\item{data}{data.frame the original, non-imputed dataset (mids$data)}

\item{imp}{list of data.frames all imputations stored in the mids object}

\item{j}{character scalar the name of the variable whose imputations should be replaced}

\item{i}{character or integer scalar the number of the current imputation (can be 1:m)}
}
\description{
{ Replace all overimputed data points in the mice imputation
of one variable. Overimputed data points are those data
that were not missing in the original but were marked for
imputation manually and imputed by the imputation procedure.
}
}
