% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_set_Pipeline.R
\name{single_set_Pipeline}
\alias{single_set_Pipeline}
\title{Pipline for one data set}
\usage{
single_set_Pipeline(
  df,
  id,
  col_type,
  percent_of_missing,
  out_file_location,
  single_set = TRUE,
  cores = NULL
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without target column.}

\item{col_type}{character vector. Vector containing column type names.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{out_file_location}{Location of output file.}

\item{single_set}{Set as False if Pipeline is used in a bigger pipeline for multiple datasets.}

\item{cores}{Number of threads used by missForest.}

\item{in}{openml dataset id.}

\item{OTHERS}{Function will be containing much more params responsible for settings for now all no essential parameters are set on default.}
}
\value{
Return list of data.frame or NULLs if the imputation method didn't work.
}
\description{
Function is Work in progress.
}
\details{
---
}
