% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA-data.R
\docType{data}
\name{universe}
\alias{universe}
\title{A sample \verb{character vector} containing the background gene list used to
perform FEA on the \code{macrophage} dataset}
\format{
A \verb{character vector}
}
\value{
A sample \verb{character vector} containing the assay's \code{rownames}
of the \code{macrophage} data
}
\description{
A sample \verb{character vector} containing the background gene list used to
perform FEA on the \code{macrophage} dataset
}
\details{
This \verb{character vector} object contains the assay's \code{rownames}
of the \code{macrophage} data

The code to create said object can be found in the folder \verb{/inst/scripts} in
the EMMA package, the file is called \code{create_datasets_examples.R}.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
