% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_missForest.R
\name{autotune_missForest}
\alias{autotune_missForest}
\title{Perform imputation using missForest form missForest package.}
\usage{
autotune_missForest(
  df,
  col_type,
  percent_of_missing,
  cores = NULL,
  ntree_set = c(100, 200, 500, 1000),
  mtry_set = NULL,
  parallel = TRUE,
  col_0_1 = FALSE,
  optimize = TRUE,
  ntree = 100,
  mtry = NULL,
  verbose = FALSE,
  maxiter = 20,
  maxnodes = NULL,
  out_file = NULL
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{cores}{integer.  Number of threads used by parallel calculations. By default approximately half of available CPU cores.}

\item{ntree_set}{integer vector. Vector contains numbers of tree for grid search.}

\item{mtry_set}{integer vector. Vector contains numbers of variables randomly sampled at each split.}

\item{parallel}{logical. If TRUE parallel calculation is using.}

\item{col_0_1}{decide if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False.}

\item{optimize}{optimize inside function}

\item{ntree}{ntree from missForest function}

\item{mtry}{mtry form missforest function}

\item{verbose}{If FALSE funtion didn't print on console.}

\item{maxiter}{maxiter form missForest function.}

\item{maxnodes}{maxnodes from missForest function.}

\item{out_file}{Output log file location if file already exists log message will be added. If NULL no log will be produced.}
}
\value{
Return data.frame with imputed values.
}
\description{
Function use missForest package for data imputation. OBBerror (more in  \code{\link{autotune_mice}}) is used to perform grid search.
}
\details{
Function try to use parallel backend if it's possible. Half of the available cores are used or number pass as cores param. (Number of used cores can't be higher then number of variables in df. If it happened a number of cores will be set at ncol(df)-2 unless this number is <= 0 then cores =1).  To perform parallel calculation function use  \code{\link[doParallel]{registerDoParallel}} to create parallel backend.
Creating backend can have significant time cost so for very small df cores=1 can speed up calculation. After calculation function turns off parallel backend. \cr \cr   Gride search is used to chose a sample for each tree and the number of trees can be turn off. Params in grid search have significant influence on imputation quality but function should work on any reasonable values of this parameter.
}
