% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_VIM_kNN.R
\name{autotune_VIM_kNN}
\alias{autotune_VIM_kNN}
\title{K nearest neighbor imputation using VIM package.}
\usage{
autotune_VIM_kNN(
  df,
  percent_of_missing,
  k = 5,
  numFun = median,
  catFun = VIM::maxCat,
  col_0_1 = FALSE,
  out_file = NULL
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without  target column.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{k}{Value of k use if optimize=FALSE}

\item{col_0_1}{decide if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False.}

\item{out_file}{Output log file location if file already exists log message will be added. If NULL no log will be produced.}

\item{numFUN}{function for aggregating the k Nearest Neighbours in the case of a numerical variable. Defoult median.}

\item{catFUN}{function for aggregating the k Nearest Neighbours in the case of a categorical variable. Defoult mode.}
}
\value{
Return data.frame with imputed values.
}
\description{
Function perform kNN function from VIM packge. Only K is tuned by function.
}
