% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_VIM_regrImp.R
\name{autotune_VIM_regrImp}
\alias{autotune_VIM_regrImp}
\title{Perform imputation using VIM package and regressionImp function.}
\usage{
autotune_VIM_regrImp(
  df,
  col_type,
  percent_of_missing,
  col_0_1 = F,
  robust = F,
  mod_cat = F,
  use_imputed = F,
  out_file = NULL
)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without target column.}

\item{col_type}{Character vector with types of columns.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}

\item{col_0_1}{Decaid if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed. Default False. (Works only for returning one dataset).}

\item{robust}{TRUE/FALSE if robust regression should be used.}

\item{mod_cat}{TRUE/FALSE if TRUE for categorical variables the level with the highest prediction probability is selected, otherwise it is sampled according to the probabilities.}

\item{use_imputed}{TRUE/FALSe if TURE already imputed columns will be used to impute another.}

\item{out_file}{Output log file location if file already exists log message will be added. If NULL no log will be produced.}
}
\value{
Return one data.frame with imputed values.
}
\description{
Function use Regression models to impute missing data.
}
