% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_add_custom_metadata.R
\name{EMMA_add_custom_metadata}
\alias{EMMA_add_custom_metadata}
\title{EMMA_add_custom_metadata}
\usage{
EMMA_add_custom_metadata(res, extra = list())
}
\arguments{
\item{res}{A functional enrichment analysis results object as returned by
\code{EMMA_run()}}

\item{extra}{A named list of user-defined metadata elements to store in
the \code{extra} field}
}
\value{
The input result object with updated \code{EMMA_record} attribute
}
\description{
Append or replace the \code{extra} field in the \code{EMMA_record} attribute
of a result object returned by \code{EMMA_run()}. This allows users to manually
provide additional annotation or contextual information that could not be
captured automatically
}
\examples{
data("de_res_IFNg_vs_naive", package = "EMMA")
data("universe", package = "EMMA")
library("clusterProfiler")
res <- EMMA_run(
  enrichGO(gene = rownames(de_res_IFNg_vs_naive),
    universe = universe, keyType = "ENSEMBL",
    OrgDb = org.Hs.eg.db::org.Hs.eg.db,
    ont = "BP"
  )
)
res <- EMMA_add_custom_metadata(
  res,
  extra = list(note = "The background gene set list was all expressed genes in the assay")
)
}
