% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_explain.R
\name{EMMA_explain}
\alias{EMMA_explain}
\title{EMMA_explain}
\usage{
EMMA_explain(res)
}
\arguments{
\item{res}{A functional enrichment analysis results object as returned by
\code{EMMA_run()}. Its attributes contain \code{EMMA_record} of class \code{EMMARecord}, which
contains all provenance information of the performed FEA}
}
\description{
This function generates a human-readable description of the FEA, similar
to a Materials and Methods section of a paper, by summarizing the executed
call, the parameters, software context, and reference databases used.
}
