% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_explain.R
\name{EMMA_explain}
\alias{EMMA_explain}
\title{EMMA_explain}
\usage{
EMMA_explain(res, get_citation = TRUE)
}
\arguments{
\item{res}{A functional enrichment analysis results object as returned by
\code{EMMA_run()}. Its attributes contain \code{EMMA_record}, which contains all
provenance information of the performed FEA}

\item{get_citation}{Logical indicating whether to display the citations of
the packages used in the FEA. It only prints the citations in an interactive
session (e.g console). Defaults to \code{TRUE}}
}
\value{
A character string describing how the FEA was performed using the
recorded metadata
}
\description{
This function generates a human-readable description of the FEA, similar
to a Materials and Methods section of a paper, by summarizing the executed
call, the parameters, software context, and reference databases used.
}
\examples{
data("fea_res", package = "EMMA")
EMMA_explain(fea_res)

}
