% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_freeze.R
\name{EMMA_freeze}
\alias{EMMA_freeze}
\title{EMMA_freeze}
\usage{
EMMA_freeze()
}
\description{
This function records the R environment during analysis runtime and generates
a lockfile that can be used with tools such as \code{renv}
}
