% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_get_record.R
\name{EMMA_get_record}
\alias{EMMA_get_record}
\title{EMMA_get_record}
\usage{
EMMA_get_record(res)
}
\arguments{
\item{res}{Functional Enrichment Analysis results (enrichResult, gseaResult ...)
generated by \code{EMMA_run()}}
}
\value{
list of metadata recorded during FEA runtime
}
\description{
EMMA_get_record
}
\examples{
data("fea_res", package = "EMMA")
EMMA_get_record(fea_res)
}
\seealso{
\code{\link[clusterProfiler:enrichGO]{enrichGO()}}, \code{\link[clusterProfiler:groupGO]{groupGO()}}, \code{\link[clusterProfiler:gseGO]{gseGO()}}, \code{\link[clusterProfiler:gseKEGG]{gseKEGG()}}, \code{\link[clusterProfiler:enrichKEGG]{enrichKEGG()}},
\code{\link[gprofiler2:gost]{gost()}}, \code{\link[mosdef:run_cluPro]{run_cluPro()}}, \code{\link[mosdef:run_topGO]{run_topGO()}}, \code{\link[mosdef:run_goseq]{run_goseq()}}
}
