% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_run.R
\name{EMMA_run}
\alias{EMMA_run}
\title{EMMA_run}
\usage{
EMMA_run(
  expr,
  envir = parent.frame(),
  session = TRUE,
  args_form = c("evaluated", "unevaluated")
)
}
\arguments{
\item{expr}{A function call that performs functional enrichment analysis.
The call is captured and executed by EMMA to record analysis parameters and
provenance information. Both bare calls (\code{enrichGO(...)}) and namespace-qualified
calls (\code{clusterProfiler::enrichGO(...)}) are supported. Any other form
(e.g. \code{do.call}, \code{get()}) will raise an error}

\item{envir}{An environment in which to evaluate \code{expr}}

\item{session}{Logical, indicating whether to store the output of
\code{sessionInfo()} or not. If \code{TRUE} (default), the session is stored in the
provenance record}

\item{args_form}{A character string indicating whether to store the evaluated
or the unevaluated arguments in the provenance record. It default to \code{"evaluated"}}
}
\value{
The result object returned by the enrichment function in \code{expr},
unmodified except for an added \code{EMMA_record} attribute containing the
provenance information. Use \code{getEMMARecord()} to access it
}
\description{
This function executes any supported functional enrichment analysis function
and automatically captures the call, its associated parameters and provenance
information when available during runtime as an \code{EMMA_record} attribute
on the returned results object.
}
\examples{
data("de_res_IFNg_vs_naive", package = "EMMA")
data("universe", package = "EMMA")
EMMA_run(clusterProfiler::enrichGO(gene = rownames(de_res_IFNg_vs_naive),
universe = universe, keyType = "ENSEMBL", OrgDb = org.Hs.eg.db::org.Hs.eg.db,
ont = "BP"))
}
