% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_show.R
\name{EMMA_show}
\alias{EMMA_show}
\title{EMMA_show}
\usage{
EMMA_show(res)
}
\arguments{
\item{res}{A functional enrichment analysis results object as returned by
\code{EMMA_run()}}
}
\value{
A list of the recorded information
}
\description{
EMMA_show
}
\examples{
res <- EMMA_run(enrichGO(gene = rownames(de_res_IFNg_vs_naive), universe = universe, keyType = "ENSEMBL", OrgDb = org.Hs.eg.db, ont = "BP"))
EMMA_show(res)
}
