% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA_show.R
\name{EMMA_show}
\alias{EMMA_show}
\title{EMMA_show}
\usage{
EMMA_show(res)
}
\arguments{
\item{res}{A functional enrichment analysis results object as returned by
\code{EMMA_run()}}
}
\value{
\code{base::invisible()}
}
\description{
This function displays a human-readable summary of the \code{EMMA_record} attribute
attached to a result object produced by \code{EMMA_run()}
}
\examples{
data("fea_res", package = "EMMA")
EMMA_show(fea_res)
}
