% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_mice.R
\name{formula_creating}
\alias{formula_creating}
\title{Creating formula}
\usage{
formula_creating(df,coll_miss,coll_no_miss,coll_type,percent_of_missing)
}
\arguments{
\item{df}{data.frame. Df to impute with column names and without target column.}

\item{col_miss}{character vector. Names of columns with NA.}

\item{col_no_miss}{character vector. Names of columns without NA.}

\item{col_type}{character vector. Vector containing column type names.}

\item{percent_of_missing}{numeric vector. Vector contatining percent of missing data in columns for example  c(0,1,0,0,11.3,..)}
}
\value{
List with formula object[1] and information if its no numeric value in dataset[2].
}
\description{
This function create formula to use in with function form mice package.
Also inform if its necessary to use gml instead of lm (no numeric values in dataset).
}
