% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMMARecord.R
\name{getEMMARecord}
\alias{getEMMARecord}
\title{getEMMARecord}
\usage{
getEMMARecord(res)
}
\arguments{
\item{res}{Functional Enrichment Analysis results (enrichResult, gseaResult ...)}
}
\value{
list of metadata recorded during FEA runtime
}
\description{
getEMMARecord
}
\examples{
data("de_res_IFNg_vs_naive", package = "EMMA")
data("universe", package = "EMMA")
library("clusterProfiler")
res <- EMMA_run(enrichGO(gene = rownames(de_res_IFNg_vs_naive),
universe = universe, keyType = "ENSEMBL", OrgDb = org.Hs.eg.db::org.Hs.eg.db,
ont = "BP"))
getEMMARecord(res)
}
