% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpMice.R
\name{PipeOpMice}
\alias{PipeOpMice}
\title{PipeOpMice}
\description{
Implements mice methods as mlr3 pipeline more about mice \code{\link{autotune_mice}}
}
\section{Input and Output Channels}{

Input and output channels are inherited from \code{\link{PipeOpImpute}}.
}

\section{Parameters}{

The parameters are the parameters inherited from [`PipeOpImpute`], as well as: \cr
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier of resulting object, default \code{"imput_mice"}.
\item \code{m} :: \code{integer(1)}\cr
number of sets produced by mice, default \code{5}.
\item \code{maxit} :: \code{integer(1)}\cr
maximum number of iteration for mice, default \code{5}.
\item \code{set_corr} :: \code{double(1)}\cr
Correlation or fraction of featurs using if optimize= False. When correlation == F  its represent a fraction of features to use in imputation for each variable.  default \code{0.5}.
\item \code{set_method} :: \code{character(1)}\cr
Method used if optimize=False. If NULL default method is used (more in methods_random section ), default \code{'pmm'}.
\item \code{low_corr} :: \code{double(1)}\cr
double betwen 0,1 default 0 lower boundry of correlation used in inner optimization (used only when optimize == TRUE), default \code{0}.
\item \code{up_corr} :: \code{double(1)}\cr
double between 0,1 default 1 upper boundary of correlation set used in inner optimization (used only when optimize == TRUE). Both of these parameters work the same for a fraction of features if correlation == FALSE.,defoult \code{1}.
\item \code{methods_random} :: \code{character(1)}\cr
set of methods to chose. Default 'pmm'. If seted on NULL this methods are used predictive mean matching (numeric data) logreg, logistic regression imputation (binary data, factor with 2 levels) polyreg, polytomous regression imputation for unordered categorical data (factor > 2 levels) polr, proportional odds model for (ordered, > 2 levels).
\item \code{iter} :: \code{integer(1)}\cr
number of iteration for randomSearch, default \code{5}.
\item \code{random.seed} :: \code{integer(1)}\cr
random seed, default \code{123}.
\item \code{optimize} :: \code{logical(1)}\cr
If set TRUE function will optimize parametrs of imputation automaticlly. If parametrs will be tune by other methode shoude be set as FALSE, default \code{FALSE}.
\item \code{correlation} :: \code{logical(1)}\cr
If True correlation is using if Fales fraction of features, default \code{TRUE}.
\item \code{col_0_1} :: \code{logical(1)}\cr
 if add bonus column informing where imputation been done. 0 - value was in dataset, 1 - value was imputed, Default \code{FALSE}.
\item \code{out_fill} :: \code{character(1)}\cr
Output log file location if file already exists log message will be added. If NULL no log will be produced, default \code{NULL}.
}
}

\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpImpute]{mlr3pipelines::PipeOpImpute}} -> \code{mice_imputation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PipeOpMice$new()}}
\item \href{#method-clone}{\code{PipeOpMice$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict">}\href{../../mlr3pipelines/html/PipeOp.html#method-predict}{\code{mlr3pipelines::PipeOp$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print">}\href{../../mlr3pipelines/html/PipeOp.html#method-print}{\code{mlr3pipelines::PipeOp$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train">}\href{../../mlr3pipelines/html/PipeOp.html#method-train}{\code{mlr3pipelines::PipeOp$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpMice$new(
  id = "imput_mice",
  m = 5,
  maxit = 5,
  set_cor = 0.5,
  set_method = "pmm",
  low_corr = 0,
  up_corr = 1,
  methods_random = c("pmm"),
  iter = 5,
  random.seed = 123,
  optimize = F,
  correlation = F,
  col_0_1 = F,
  out_file = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpMice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
