% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpVIM_kNN.R
\name{PipeOpVIM_kNN}
\alias{PipeOpVIM_kNN}
\title{VIM kNN Imputation}
\description{
This class create object implements autotune_VIM_kNN function for use in mlr3 pipelinies. Object can be created with \code{\link{autotune_VIM_kNN}} params.
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpImpute]{mlr3pipelines::PipeOpImpute}} -> \code{VIM_kNN_imputation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PipeOpVIM_kNN$new()}}
\item \href{#method-predict_internal}{\code{PipeOpVIM_kNN$predict_internal()}}
\item \href{#method-train_internal}{\code{PipeOpVIM_kNN$train_internal()}}
\item \href{#method-clone}{\code{PipeOpVIM_kNN$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict">}\href{../../mlr3pipelines/html/PipeOp.html#method-predict}{\code{mlr3pipelines::PipeOp$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print">}\href{../../mlr3pipelines/html/PipeOp.html#method-print}{\code{mlr3pipelines::PipeOp$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train">}\href{../../mlr3pipelines/html/PipeOp.html#method-train}{\code{mlr3pipelines::PipeOp$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOpImpute" data-id="impute">}\href{../../mlr3pipelines/html/PipeOpImpute.html#method-impute}{\code{mlr3pipelines::PipeOpImpute$impute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOpImpute" data-id="select_cols">}\href{../../mlr3pipelines/html/PipeOpImpute.html#method-select_cols}{\code{mlr3pipelines::PipeOpImpute$select_cols()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOpImpute" data-id="train_imputer">}\href{../../mlr3pipelines/html/PipeOpImpute.html#method-train_imputer}{\code{mlr3pipelines::PipeOpImpute$train_imputer()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpVIM_kNN$new(
  id = "imput_VIM_kNN",
  k = 5,
  numFun = median,
  catFun = maxCat,
  col_0_1 = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_internal"></a>}}
\if{latex}{\out{\hypertarget{method-predict_internal}{}}}
\subsection{Method \code{predict_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpVIM_kNN$predict_internal(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_internal"></a>}}
\if{latex}{\out{\hypertarget{method-train_internal}{}}}
\subsection{Method \code{train_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpVIM_kNN$train_internal(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpVIM_kNN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
